feature(component): Adds component decorators
Adds decorators for tagName, classNames, classNameBindings, and
attributeBindings. These fields, for a variety of reasons, have special
behavior when compared to other class fields such as merging, concatenating
or simply having to be applied to the prototype/before init. These decorators
provide easy to use, declarative alternatives which should tide us over
until we are able to drop them altogether (with Glimmer components)