bugfix(props): Normalizes property configurations by default (#171)
This fixes an issue in the babel transform for decorators which is a
longstandanding bug. When class fields are decorated, they do not provide
`writable` or `configurable` values. Instead, `writable` is set by
`_applyDecoratedDescriptor` like so:

```js
  if ('value' in desc || desc.initializer) {
    desc.writable = true;
  }

  desc = decorators.slice().reverse().reduce(function (desc, decorator) {
    return decorator(target, property, desc) || desc;
  }, desc);
```

The issue arises when we define a class field without an initializer, i.e.

```js
class Foo() {
  @service bar;
}
```

The initial value is set to `null`, which means the initializer is falsy,
so the descriptor is not modified. By the time it gets to our decorators
the descriptor's `writable` value is still unset, meaning that a class field
which likely should be writable is not.

This specifically clashes with legacy injections, which can cause the
container to inject a field twice. It has popped up in other cases as
well, such as the `@className` and `@attribute` decorators. Considering
that class methods are by default writable, it makes sense that fields
should be too unless they are otherwise specified (via another decorator,
for instance).