[bugfix] Do not use inherited descriptors (#230)
Currently, computed property decorators receive the last CP descriptor that
was assigned to the property. This way, a decorator can build off the last
decorator. This functionality is only used by the @readOnly decorator, which
marks an existing computed property as read only. The decorator is not
particularly useful anymore since for most standard computed properties,
all you need to do is not add a setter, so rather than keeping this behavior
this PR deprecates it and marks it for removal.

In the meantime, we still need to get the descriptor for the exact class
we're decorating - otherwise, we may attempt to decorate the superclass
instead, which was the root cause of #214