[bugfix] Always collapse proto from constructor
Currently we check the super class to see if it has `proto` and call it,
but that changed with @krisselden's recent PRs. Now only the base class
actually has `proto`, and class that inherits does get the method.

I believe the original implementation was overly cautious, calling proto
on subclasses should be fine all the way back through much older
versions of Ember. Everything should get applied correctly, because the
proto method only ever operated on the class in the closure until now,
it didn't use `this` or any context.