[feat] Adds computed modifiers
Adds the following modifiers for computed properties:

```js
@computed.readOnly()
@computed.volatile()
@(computed().property())
@(computed().meta())
```

These modifiers map to the built in modifiers in computed properties,
and are primarily being added to facilitate codemodding existing code.
Documentation has only been added for `.readOnly`, as the other
modifiers are _not_ recommended for use outside of updating existing
code.

This PR also removes the deprecation warning provided by the setter in
debug mode, unfortunately we cannot provide a setter _and_ allow users
to call `readOnly`. We'll have to wait for the official deprecation to
inform users that `readOnly` has been deprecated, if it comes along.