[REFACTOR] Rename readOnly -> reads, reads/oneWay -> overridableReads (#283)
The unfortunate naming conflict between the two different kinds of
`@readOnly` decorators is very confusing and makes it difficult to tell
what is going on without context. Now that overridability has been
deprecated, the naming makes even less sense.

This PR renames the `@readOnly` decorator to `@reads`, since it will
represent the default behavior in the new world (non-overridable one
directional alias). It also renames `@oneWay` to be `@overridableReads`
for clarity. This macro will likely continue to exist even in the
post-`readOnly` world, and continue to have the same functionality it
has now, so it makes sense to name it appropriately.