[FEAT] Stage 2 Decorators
Adds support for the stage 2 decorator API, while maintaining support
for the v1 API via a compatibility layer. If no usages of the v1
decorators are detected in the build, stage 1 code is removed entirely
in favor of the stage 2 code, resulting in a lighter and more performant
build. Usage of the stage 1 transforms is deprecated and users will be
warned, and the offending package will be labeled so it can be updated.