[FEAT] Stage 2 transforms and support (#304)
* Build updates

* Update and pin ember-cli-babel to v7
* Update to the stage 2 transforms in babel plugins
* Update types and ember-cli-typescript to v2
* Remove custom TS test generation in favor of a renaming funnel
* Add ember-try scenarios for stage 1 transforms
* Remove `javascript` from test names now that TS and JS are not tested
  side-by-side

* tests

* [FEAT] Stage 2 Decorators

Adds support for the stage 2 decorator API, while maintaining support
for the v1 API via a compatibility layer. If no usages of the v1
decorators are detected in the build, stage 1 code is removed entirely
in favor of the stage 2 code, resulting in a lighter and more performant
build. Usage of the stage 1 transforms is deprecated and users will be
warned, and the offending package will be labeled so it can be updated.

* adds build checks for stage 1