[BUGFIX] Fixes usage of `super` within actions
Currently, in a pure-native-class world `super` will work exactly as
expected in actions, because actions are just methods now. However, when
actions were defined using the _traditional_ class model, they only
exist on the `actions` hash, and there is no easy way to call the super
action.

This ensures that `@action` checks for calls to super, and if they exist
adds a shim method which calls the super action. The shim is only added
if one does not already exist, and if it would not conflict with another
value. If it would conflict, an error is thrown.

It should be noted that calling `super` on a normal super method _will_
still work, since this follows the normal rules of super in native
classes. The new paradigm, that `@action` just annotates a method, is
preserved.