fix(utils/decorator): transfer initializer, detect synthetic descriptors (#318)
If we are dealing with legacy decorators and a decorated member has an `initializer`, the `initializer` is kept on the property descriptor and not transferred to the synthetic member descriptor. This PR deletes the `initializer` from the original descriptor and transfers it.

dd9348a0dca4e43af157d645f202bc92db544ce8 then transfers the initializer back to the property descriptor, just before we give it back to Babel.

Relevant part of the spec: [Field Decorator > Parameters](https://github.com/tc39/proposal-decorators/blob/master/METAPROGRAMMING.md#parameters)

This PR also adds a `toString` method to the synthetic member descriptor, so that `isStage2Descriptor` detects synthetic member descriptors as such to enable the pattern shown in https://github.com/ember-decorators/ember-decorators/issues/315#issuecomment-441720711 for legacy decorators as well.