[BUGFIX] Checks devDeps for ec-typescript
`ember-cli-typescript` v1 would precompile addons _before_ they were
published, meaning it would be in `devDependencies` and
ember-cli-version-checker would not be able to pick up on them. This
adds a check to dependencies/devDependencies for `ember-cli-typescript`,
which should solve a lot of the larger ecosystem bugs that have been
popping up.