[BUGFIX] Move all types to proper location in the root (#373)
The recent change to move types into the `/published-types` directory and use the `"types"` field of `package.json` breaks down for `@ember-decorators/object` which has more than one module. [Limitations in the way that package.json's `main` and `types` fields](https://github.com/Microsoft/TypeScript/issues/8305) work prevent us from being able to point to a root directory, we can only point to one file for the index, so this strategy doesn't work for us.

This PR switches us back to using `prepack` and `postpack` steps to move the typings to the correct location. This is the same direction that Embroider is moving too, so I'm pretty comfortable with doing this until that effort has been completed and we can rely on more standardized pre-publish tooling.

Fixes #372, cc @buschtoens @mike-north