[BUGFIX] Fixes observers (#383)
Fixes two major problems with observers:

1. They didn't expand properties
2. They would completely break if used with non EmberObject based
   classes.

This PR fixes both issues, but only for stage 2 transforms. It works by
adding another field to the class that is a dummy field and finishes
chains during initialization.

This is very very hacky, but it should be possible to do this much
better with `kind = 'initializer'` descriptors in stage 3, once they're
implemented. It shouldn't mess with shape or anything (I don't think),
will just be a bit ugly. The prop is non-enumerable too, so should be
very little impact.